refactor: steps modules import downward only, no deferred imports
Round-0 broke two circular imports (_loops<->_loop_steps,
project_registration<->_project_registration_steps) with deferred
function-body imports. That resolves the cycle at runtime but the cycle is
still there structurally, which check_imports.py cannot see since it only
flags cross-layer imports.

Cycle A: moved _loop_should_stop, _retry_dim_callback, _finalize_dim_result,
and _dispatch_incremental_dim into _loop_steps.py, next to
_run_one_incremental_dim. None of them need anything from _loops.py.
_loops.py now imports _loop_steps at the top; run_per_dimension_loop, which
still calls _finalize_dim_result and _loop_should_stop, imports them back
from there too.

Cycle B: moved _resolve_target_path, _persist_repository_info,
_ensure_onboarding_field, and _LOCATION_LOCAL into
_project_registration_steps.py. They had no remaining callers in
project_registration.py. _validate_clone_target and
mark_onboarding_complete stay put since they're unrelated to what moved.
project_registration.py now imports the steps module at the top.

Retargeted every test patch on a name whose call site moved (13 occurrences
across 5 files). Two tests exercise validate_remote_url from both the
guard that stayed and the pre-clone check that moved, so those now patch
both module paths. Regenerated the param baseline: three already-
grandfathered, unchanged-signature functions changed file, so their
baseline keys moved with them.

Also dropped _cancel_running's snapshot parameter, it was never read.
