+++ src/quodeq/services/_project_registration_steps.py
@@ -2,30 +2,113 @@
 materialize it on disk and scan.
 
 Split out of ``project_registration.py`` (Task 22, M-MOD-6): that module sits
-at the 300-line size ratchet, so these steps live here instead. Imports back
-from ``project_registration`` for the callees it absorbs (``_resolve_target_path``,
-``_persist_repository_info``, ``_ensure_onboarding_field``): those stay put
-because tests patch the underlying I/O (``run_git_clone``, ``validate_remote_url``,
-``write_repository_info``) by ``quodeq.services.project_registration.<name>``,
-which only resolves correctly if the callers of those names stay defined
-there. ``register_project`` imports this module back with a deferred import
-to avoid a circular top-level import between the two.
+at the 300-line size ratchet, so these steps live here instead. This module
+imports only downward (``_wiring``, ``_fs_clone``, ``_fs_scan``,
+``_registration_scan``, ``_registration_url``, ``shared``) and never imports
+back from ``project_registration`` -- ``project_registration.py`` imports it
+at the top instead. ``_resolve_target_path``, ``_persist_repository_info``,
+and ``_ensure_onboarding_field`` moved here too (they had no remaining
+callers left in ``project_registration.py``); tests that patch the I/O they
+call (``run_git_clone``, ``validate_remote_url``, ``write_repository_info``)
+target ``quodeq.services._project_registration_steps.<name>`` now.
 """
 from __future__ import annotations
 
 from dataclasses import dataclass
 from pathlib import Path
 
+from quodeq.services._fs_clone import run_git_clone
 from quodeq.services._fs_scan import scan_project
 from quodeq.services._registration_scan import _scan_parent_project
-from quodeq.services._wiring import ProjectIdentity, resolve_project_uuid
-from quodeq.services.project_registration import (
-    _LOCATION_LOCAL,
-    _ensure_onboarding_field,
-    _persist_repository_info,
-    _resolve_target_path,
+from quodeq.services._registration_url import _read_origin_remote, _strip_credentials
+from quodeq.services._wiring import (
+    ProjectIdentity,
+    read_repository_info,
+    resolve_project_uuid,
+    validate_remote_url,
+    write_repository_info,
 )
+from quodeq.shared._env import get_clones_dir
 from quodeq.shared.utils import is_repo_url, project_name_from_repo
+
+_LOCATION_LOCAL = "local"
+
+
+def _resolve_target_path(
+    repo: str, repo_resolved: str, project_name: str, project_uuid: str, *,
+    is_url: bool, ephemeral: bool, clone_dest: str | None, clones_dir: Path | None,
+) -> Path:
+    """Resolve/create the on-disk path the project will live at.
+
+    For a URL input, clones into an ephemeral cache dir or the caller's
+    chosen *clone_dest*. For a local path input, resolves in place -- the
+    directory must already exist.
+    """
+    if is_url:
+        if ephemeral:
+            target_path = (clones_dir or get_clones_dir()) / project_uuid
+        else:
+            target_path = Path(clone_dest).resolve() / project_name
+        target_path.parent.mkdir(parents=True, exist_ok=True)
+        # Re-validate immediately before the clone dispatch: the project-uuid
+        # resolution between _validate_clone_target's check and here (index
+        # load, legacy directory scan, project creation) does real disk I/O and
+        # can take enough wall-clock time for a DNS-rebinding attacker to flip
+        # the host from a public to a private IP. This narrows, but does not
+        # close, the race -- git clone re-resolves DNS again itself, independently,
+        # inside the subprocess below; only pinning the resolved IP through git's
+        # own connection (a hosts-file override or proxy layer) would close it,
+        # and that's out of scope here.
+        validate_remote_url(repo)
+        # run_git_clone raises CloneError on failure (Task A8). We let it propagate.
+        run_git_clone(repo, target_path)
+        return target_path
+
+    target_path = Path(repo_resolved)
+    if not target_path.is_dir():
+        # A path pointing at a FILE is a distinct user mistake from a
+        # missing path (a real registration once slipped through as
+        # .../lib/player.js) — say which one it was.
+        detail = "points at a file, not a directory" if target_path.exists() else "does not exist"
+        raise FileNotFoundError(f"Repo path {detail}: {target_path}")
+    return target_path
+
+
+def _persist_repository_info(
+    project_dir: Path, target_path: Path, *, is_url: bool, repo: str, ephemeral: bool,
+) -> None:
+    """Persist the resolved path + ephemeral flag in repository_info.json.
+
+    A corrupt existing file is treated as empty and rewritten (self-heal);
+    registration is the flow that owns this file's creation.
+    """
+    info = read_repository_info(project_dir) or {}
+    info["path"] = str(target_path.resolve())
+    info["location"] = _LOCATION_LOCAL
+    info["ephemeral"] = bool(ephemeral)
+    origin_url = repo if is_url else _read_origin_remote(target_path)
+    if origin_url:
+        # Defense in depth: _read_origin_remote already strips credentials
+        # from the local-remote branch, but strip again here so the
+        # URL-registration branch (raw *repo*) is covered too, and so this
+        # call site stays safe even if the helper's behavior changes.
+        info["originUrl"] = _strip_credentials(origin_url)
+    write_repository_info(project_dir, info)
+
+
+def _ensure_onboarding_field(project_dir: Path) -> None:
+    """Add `onboardingCompletedAt: null` to repository_info.json if absent.
+
+    Called during registration (via `_resolve_project_slot`) so newly-registered
+    projects start with the field set to null. Existing projects without the
+    field get a backfill on read (see `_backfill_onboarding_field` in
+    _fs_project_helpers.py).
+    """
+    data = read_repository_info(project_dir)
+    if data is None or "onboardingCompletedAt" in data:
+        return
+    data["onboardingCompletedAt"] = None
+    write_repository_info(project_dir, data)
 
 
 def _resolve_project_slot(
